requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Celery==5.3.4
Redis==5.0.1
SQLAlchemy==2.0.23
Werkzeug==2.3.7
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
pandas==2.1.4
email-validator==2.1.0
cryptography>=41.0.0
gunicorn==21.2.0
Pillow==10.1.0
reportlab==4.0.7
pytz==2023.3
